% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.apply_formats}
\alias{protobi.apply_formats}
\title{Apply Formats Function}
\usage{
protobi.apply_formats(data_df, format_df)
}
\arguments{
\item{data_df}{#R dataframe}

\item{format_df}{#format metadata}
}
\description{
Given the data and format, this function replaces the values in the dataframe column with the format metada values.
}
\keyword{protobi}
\keyword{protobi.get_formats()}
