% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.apply_titles}
\alias{protobi.apply_titles}
\title{Apply Titles Function}
\usage{
protobi.apply_titles(data_df, names_df)
}
\arguments{
\item{data_df}{#R dataframe}

\item{names_df}{#format metadata}
}
\description{
Given the data and format, this function assigns the tile (label) metadata to the R dataframe columns.
}
\keyword{protobi}
\keyword{protobi.get_titles()}
