% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.get_data}
\alias{protobi.get_data}
\title{Get Data Function}
\usage{
protobi.get_data(projectid, tablekey, apikey)
}
\description{
This function returns an R data frame representing the data in Protobi based on the parameters provided.
}
\keyword{protobi}
\keyword{protobi.get_data()}
