% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.get_formats}
\alias{protobi.get_formats}
\title{Get Formats Function}
\usage{
protobi.get_formats(projectid, apikey)
}
\description{
This function returns an R List representing the Format metadata in Protobi based on the parameters provided.
}
\keyword{protobi}
\keyword{protobi.get_formats()}
