% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.get_titles}
\alias{protobi.get_titles}
\title{Get Titles Function}
\usage{
protobi.get_titles(projectid, apikey)
}
\description{
This function returns an R List representing the Titles metadata in Protobi based on the parameters provided.
}
\keyword{protobi}
\keyword{protobi.get_titles()}
