% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_apply_formats}
\alias{protobi_apply_formats}
\title{Apply Formats Function}
\usage{
protobi_apply_formats(data_df, format_df)
}
\arguments{
\item{data_df}{A dataframe.}

\item{format_df}{A format metadata.}
}
\value{
The data_df with levels adjusted according to format_df.
}
\description{
Given the data and format, this function replaces the values in the dataframe column with the format metadata values.
}
\keyword{protobi}
\keyword{protobi_get_formats()}
