% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_apply_meta}
\alias{protobi_apply_meta}
\title{Internal helper function used to apply metadata to df}
\usage{
protobi_apply_meta(df, metadata, updater)
}
\arguments{
\item{df}{A data.frame.}

\item{metadata}{A list.}

\item{updater}{A function that takes a column name and returns corresponding data with attached metadata.}
}
\value{
df with metadata attached.
}
\description{
Given df, metadata and updating function return df with updated columns
}
