% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_apply_titles}
\alias{protobi_apply_titles}
\title{Apply Titles Function}
\usage{
protobi_apply_titles(df, titles)
}
\arguments{
\item{df}{A data.frame.}

\item{titles}{A list with title metadata as returned from protobi_get_titles}
}
\value{
df with metadata attached.
}
\description{
Given the data and titles, this function assigns the tile (label) metadata to the R dataframe columns.
}
\keyword{protobi}
