% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_get_properties}
\alias{protobi_get_properties}
\title{Get properties}
\usage{
protobi_get_properties(
  projectid,
  apikey,
  execkey,
  propertykey = NULL,
  host = "https://app.protobi.com"
)
}
\arguments{
\item{projectid}{A character. Protobi project identifier.}

\item{apikey}{A character. The APIKEY from your account profile, https://app.protobi.com/account.}

\item{execkey}{A character. The key of your Protobi executable process.}

\item{propertykey}{A character. Accepts valid (not blank, NULL, or NA) non-nested lists, vectors, or character values of property key(s). If not supplied, returns all properties.}

\item{host}{URL, defaults to "https://app.protobi.com"}
}
\value{
An R list or character value object. Type dependent on properties requested.
}
\description{
protobi_get_properties() dynamically gets all supplied properties from Protobi.
}
\seealso{
[protobi_put_properties()]
}
