% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_get_titles}
\alias{protobi_get_titles}
\title{Get Titles Function}
\usage{
protobi_get_titles(projectid, apikey, host)
}
\arguments{
\item{projectid}{A character. Protobi project identifier.}

\item{apikey}{A character. e APIKEY from your account profile, https://app.protobi.com/account.}
}
\value{
A list representing titles metadata
}
\description{
This function returns an R List representing the Titles metadata in Protobi based on the parameters provided.
}
\keyword{protobi}
