% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_get_url}
\alias{protobi_get_url}
\title{Get properties directly via URL as an R object}
\usage{
protobi_get_url(url, apikey = NULL)
}
\arguments{
\item{url}{A character. URL location of requested properties.}

\item{apikey}{A character. Optional API key sent via the x-api-key header.}
}
\value{
An R list or character value object. Type dependent on properties requested.
}
\description{
protobi_get_url() returns a json-translated R object based on a user-supplied URL.
}
\seealso{
[protobi_put_url()]
}
\keyword{protobi}
