% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_put_properties}
\alias{protobi_put_properties}
\title{Upload properties}
\usage{
protobi_put_properties(
  projectid,
  apikey,
  execkey,
  propertykeyvalue,
  host = "https://app.protobi.com"
)
}
\arguments{
\item{projectid}{A character. Protobi project identifier.}

\item{apikey}{A character. The APIKEY from your account profile, https://app.protobi.com/account.}

\item{execkey}{A character. The key of your Protobi executable process.}

\item{propertykeyvalue}{A character. Accepts valid (not blank, NULL, or NA) non-nested named lists of property key.value pair(s).}

\item{host}{URL, defaults to "https://app.protobi.com"}
}
\value{
An httr response object
}
\description{
protobi_put_properties() dynamically uploads all supplied properties (as list key.value pairs) to Protobi.
}
\seealso{
[protobi_get_properties()]
}
