% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_run_process}
\alias{protobi_run_process}
\title{Run Data Process Function}
\usage{
protobi_run_process(
  projectid,
  tablekey,
  apikey,
  host = "https://app.protobi.com",
  timeout_seconds = 300
)
}
\arguments{
\item{projectid}{A character. Protobi project identifier.}

\item{tablekey}{A character. The key of your Protobi data table.}

\item{apikey}{A character. The APIKEY from your account profile, https://app.protobi.com/account.}

\item{host}{URL, defaults to "https://app.protobi.com"}

\item{timeout_seconds}{Maximum time in seconds to wait for async task completion, defaults to 300 (5 minutes)}
}
\value{
A list with elements: complete (boolean), result (data when complete=true), and possibly callback (url) or message (string) when complete=false
}
\description{
This function executes a remote data process in Protobi
}
