% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi.put_data}
\alias{protobi.put_data}
\title{Upload Data Function}
\usage{
protobi.put_data(df, projectid, tablekey, apikey,
  tmpfile = "/tmp/RData.csv", host = "https://app.protobi.com")
}
\value{
httr response object
}
\description{
This function uploads data.frame content to Protobi
}
