% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_apply_formats}
\alias{protobi_apply_formats}
\title{Apply Formats Function}
\usage{
protobi_apply_formats(df, formats)
}
\arguments{
\item{df}{A data.frame.}

\item{formats}{A list with format metadata as returned by protobi_get_formats.}
}
\value{
The df with levels adjusted according to formats.
}
\description{
Given the data and format, this function replaces the values in the dataframe column with the format metadata values.
}
\keyword{protobi}
