% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_apply_titles}
\alias{protobi_apply_titles}
\title{Apply Titles Function}
\usage{
protobi_apply_titles(data_df, names_df)
}
\arguments{
\item{data_df}{A data.frame.}

\item{names_df}{A format metadata.}
}
\value{
data_df with metadata attached.
}
\description{
Given the data and format, this function assigns the tile (label) metadata to the R dataframe columns.
}
\keyword{protobi}
\keyword{protobi_get_titles()}
