% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_put_data}
\alias{protobi_put_data}
\title{Upload Data Function}
\usage{
protobi_put_data(
  df,
  projectid,
  tablekey,
  apikey,
  host = "https://app.protobi.com",
  type = "data",
  filename = "data.csv"
)
}
\arguments{
\item{df}{A data.frame to uploaded to Protobi.}

\item{projectid}{A character. Protobi project identifier.}

\item{tablekey}{A character. The key of your Protobi data table.}

\item{apikey}{A character. The APIKEY from your account profile, https://app.protobi.com/account.}

\item{host}{URL, defaults to "https://app.protobi.com"}

\item{filename}{Optional name of file, defaults to "data.csv"}
}
\value{
An httr response object
}
\description{
This function uploads data.frame content to Protobi
}
