% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_put_url}
\alias{protobi_put_url}
\title{Upload properties directly via URL}
\usage{
protobi_put_url(url)
}
\arguments{
\item{url}{A character. URL location of requested properties.}
}
\value{
An httr response object
}
\description{
protobi_put_url() uploads a key.value pair based on a user-supplied URL.
}
\seealso{
[protobi_get_url()]
}
\keyword{protobi}
