% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobi.R
\name{protobi_read_csv_gzip}
\alias{protobi_read_csv_gzip}
\title{Utility to read zipped CSV from URL}
\usage{
protobi_read_csv_gzip(uri)
}
\arguments{
\item{uri}{Address of CSV}
}
\description{
This function reads a CSV from a URL in GZip format and returns a data frame
}
